Add comments style [TEA-84]
