+++ packages/client/src/pages/Forum/pages/Topic/components/CommentsList/CommentsList.module.scss
@@ -1,3 +1,9 @@
+.root {
+}
+
+.header {
+}
+
 .list {
 }
