+++ packages/client/src/pages/Forum/pages/Topic/components/CommentsList/CommentsList.tsx
@@ -1,21 +1,17 @@
-import { Comment } from '@workspace/shared/src/models';
+import { CommentsListProps } from './types';
+import { Empty, List } from './views';
 
 import styles from './CommentsList.module.scss';
 
-export interface CommentsListProps {
-  comments: Comment[];
-}
-
 export const CommentsList = ({ comments }: CommentsListProps) => {
   return (
-    <div className={styles.list}>
-      {comments.map(comment => (
-        <div key={comment.id} className={styles.comment}>
-          <div className={styles.header}>{comment.authorId}</div>
-          <div className={styles.content}>{comment.text}</div>
-          <div className={styles.footer}></div>
-        </div>
-      ))}
-    </div>
+    <section className={styles.root}>
+      <div className={styles.header}>
+        <h4>Комментарии</h4>
+      </div>
+
+      {comments.length === 0 && <Empty />}
+      {comments.length > 0 && <List comments={comments} />}
+    </section>
   );
 };
