+++ packages/client/src/pages/Forum/pages/Topic/components/CommentsList/types.ts
@@ -1,0 +1,5 @@
+import { Comment } from '@workspace/shared/src/models';
+
+export interface CommentsListProps {
+  comments: Comment[];
+}
