+++ packages/client/src/pages/Forum/pages/Topic/components/CommentsList/views/Empty/Empty.module.scss
@@ -1,0 +1,6 @@
+.root {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  opacity: 0.7;
+}
