+++ packages/client/src/pages/Forum/pages/Topic/components/CommentsList/views/Empty/Empty.tsx
@@ -1,0 +1,9 @@
+import styles from './Empty.module.scss';
+
+export const Empty = () => {
+  return (
+    <div className={styles.root}>
+      <p>Нет комментариев</p>
+    </div>
+  );
+};
