+++ packages/client/src/pages/Forum/pages/Topic/components/CommentsList/views/List/List.module.scss
@@ -1,0 +1,44 @@
+@import '@styles/index.scss';
+
+.root {
+  overflow-y: auto;
+  overflow-x: hidden;
+  padding: 20px 0;
+}
+
+.comment {
+  display: flex;
+  flex-direction: column;
+  background: rgba($primaryColor, 0.3);
+  padding: 20px;
+  color: $onPrimaryColor;
+  border-radius: $baseBorderRadius;
+}
+
+.comment:not(:last-of-type) {
+  margin-bottom: 20px;
+}
+
+.header {
+  display: flex;
+  justify-content: start;
+  align-items: center;
+  font-size: 12px;
+  opacity: 0.75;
+}
+
+.content {
+  display: flex;
+  padding: 8px 0;
+}
+
+.footer {
+  display: flex;
+  justify-content: start;
+  align-items: center;
+}
+
+.createdAt {
+  opacity: 0.5;
+  margin-left: 12px;
+}
