+++ packages/client/src/pages/Forum/pages/Topic/components/CommentsList/views/index.ts
@@ -1,0 +1,2 @@
+export * from './List/List';
+export * from './Empty/Empty';
